Moves RemoteFeedItem and LocalFeedItem DTO models to their own files.
