+++ FeedModule/FeedModule/FeedAPI/FeedItemsMapper.swift
@@ -23,10 +23,3 @@ internal final class FeedItemsMapper {
         return root.items
     }
 }
-
-internal struct RemoteFeedItem: Codable {
-    internal let id: UUID
-    internal let description: String?
-    internal let location: String?
-    internal let image: URL
-}
