+++ FeedModule/FeedModule/FeedAPI/RemoteFeedItem.swift
@@ -1,0 +1,15 @@
+//
+//  RemoteFeedItem.swift
+//  FeedModule
+//
+//  Created by Omran Khoja on 2/9/22.
+//
+
+import Foundation
+
+internal struct RemoteFeedItem: Codable {
+    internal let id: UUID
+    internal let description: String?
+    internal let location: String?
+    internal let image: URL
+}
