+++ FeedModule/FeedModule/FeedCache/LocalFeedItem.swift
@@ -1,0 +1,22 @@
+//
+//  LocalFeedItem.swift
+//  FeedModule
+//
+//  Created by Omran Khoja on 2/9/22.
+//
+
+import Foundation
+
+public struct LocalFeedItem: Equatable {
+    public let id: UUID
+    public let description: String?
+    public let location: String?
+    public let imageURL: URL
+    
+    public init (id: UUID, description: String?, location: String?, imageURL: URL) {
+        self.id = id
+        self.description = description
+        self.location = location
+        self.imageURL = imageURL
+    }
+}
